requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
kafka-python
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.2